refactor(ci-visibility): tests refactor
