+++ tests/ci_visibility/test_coverage_per_suite.py
@@ -9,7 +9,7 @@ from ddtrace.contrib.pytest.plugin import is_enabled
 from ddtrace.internal import compat
 from ddtrace.internal.ci_visibility import CIVisibility
 from ddtrace.internal.ci_visibility.constants import COVERAGE_TAG_NAME
-from tests.ci_visibility.test_encoder import _patch_dummy_writer
+from tests.ci_visibility.util import _patch_dummy_writer
 from tests.utils import TracerTestCase
 from tests.utils import override_env
